backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.16.1
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
